docker/cli/requirements.txt
PyYAML
python-dotenv